DON-1984 Junie assisted migration of tokens to backpack-common allowing the resources to be accessed in the configuration object
