--- Backpack/src/main/res/values/backpack.button.xml
@@ -1,7 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <declare-styleable name="BpkButtonStyle">
-        <attr name="android:fontFamily" />
-        <attr name="android:textSize" />
-    </declare-styleable>
-</resources>
