+++ backpack-common/src/main/res/drawable-nodpi/bpk_beach.xml
@@ -5,8 +5,8 @@
     android:viewportHeight="24">
     <path
         android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M13.201,6.622c0.669,1.382 1.824,5.697 1.322,8.082 -0.09,0.428 -0.552,0.64 -0.967,0.489l-2.239,-0.816 -1.698,4.666q0.195,0.145 0.396,0.312c1.007,0.84 2.225,1.451 3.536,1.517 0.976,0.048 2.034,0.11 2.777,0.144 0.405,0.02 0.692,0.421 0.571,0.808a0.08,0.08 0,0 1,-0.072 0.052H3.507A0.507,0.507 0,0 1,3 21.37c0,-0.374 0.1,-0.743 0.307,-1.054 0.944,-1.424 2.359,-2.93 4.535,-2.238l1.596,-4.384 -2.02,-0.735c-0.414,-0.15 -0.631,-0.609 -0.425,-0.996 1.148,-2.15 4.807,-4.712 6.208,-5.341" />
+        android:pathData="M12,2c-4.376,0 -9.038,4 -9.96,8.016A1.71,1.71 0,0 0,3.857 12h0.381a0.815,0.815 0,0 0,0.793 -0.62C5.703,8.642 7.745,3.39 12,2m0,0c4.255,1.39 6.297,6.642 6.969,9.38a0.815,0.815 0,0 0,0.793 0.62h0.381a1.71,1.71 0,0 0,1.817 -1.984C21.038,6 16.376,2 12,2" />
     <path
         android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M13.201,6.622c3.29,1.197 5.709,5.454 5.313,8.7 -0.107,0.87 -1.057,1.29 -1.905,0.982l-0.287,-0.104a0.65,0.65 0,0 1,-0.427 -0.679c0.238,-2.225 0.127,-6.699 -2.694,-8.9M3.538,9.872c1.783,-2.742 6.373,-4.448 9.663,-3.25 -3.576,-0.128 -6.536,3.228 -7.784,5.086a0.65,0.65 0,0 1,-0.765 0.245l-0.286,-0.105c-0.847,-0.308 -1.306,-1.24 -0.828,-1.976M19.356,7.57l0.604,0.162a0.623,0.623 0,0 1,0.446 0.759,0.62 0.62,0 0,1 -0.761,0.44l-0.605,-0.16a0.624,0.624 0,0 1,-0.446 -0.76,0.62 0.62,0 0,1 0.761,-0.44m-5.8,-1.831a2.495,2.495 0,0 1,2.65 -0.928A2.495,2.495 0,0 1,17.99 7.85c-0.083,0.316 -0.226,0.6 -0.41,0.847a10,10 0,0 0,-4.025 -2.957m5.475,-1.241a0.63,0.63 0,0 1,0.853 0.226,0.616 0.616,0 0,1 -0.223,0.848l-0.538,0.31a0.63,0.63 0,0 1,-0.855 -0.225,0.62 0.62,0 0,1 0.223,-0.849zM13.049,2.915a0.63,0.63 0,0 1,0.854 0.225l0.315,0.537a0.617,0.617 0,0 1,-0.223 0.849,0.63 0.63,0 0,1 -0.854,-0.227l-0.316,-0.537a0.617,0.617 0,0 1,0.224 -0.847m3.787,-0.5a0.624,0.624 0,0 1,0.446 0.76l-0.158,0.599a0.62,0.62 0,0 1,-0.762 0.44,0.624 0.624,0 0,1 -0.445,-0.76l0.156,-0.6a0.623,0.623 0,0 1,0.763 -0.44" />
+        android:pathData="M12,2c-1.377,1.349 -4.578,5.963 -5.008,9.004a0.896,0.896 0,0 0,0.926 0.996h8.164a0.897,0.897 0,0 0,0.926 -0.996C16.578,7.963 13.377,3.349 12,2m0,11a1,1 0,0 0,-1 1v7a1,1 0,0 0,2 0v-7a1,1 0,0 0,-1 -1" />
 </vector>
