+++ backpack-common/src/main/res/drawable-nodpi/bpk_beach_sm.xml
@@ -5,8 +5,8 @@
     android:viewportHeight="24">
     <path
         android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M12.237,7.688c0.666,1.39 1.81,5.734 1.301,8.138 -0.091,0.433 -0.553,0.646 -0.968,0.495l-2.24,-0.816 -1.677,4.612c1.098,0.787 2.35,1.355 3.7,1.355h2.071c0.53,0 0.958,0.429 0.958,0.958a0.12,0.12 0,0 1,-0.12 0.12H2.007a0.507,0.507 0,0 1,-0.507 -0.507c0,-0.374 0.1,-0.742 0.306,-1.053 1.028,-1.55 2.612,-3.197 5.129,-2.003l1.516,-4.166 -2.019,-0.733c-0.415,-0.152 -0.63,-0.613 -0.423,-1.002 1.155,-2.17 4.824,-4.761 6.228,-5.398" />
+        android:pathData="M12,1.5c-4.595,0 -9.49,4.2 -10.458,8.416A1.796,1.796 0,0 0,3.45 12h0.4a0.856,0.856 0,0 0,0.833 -0.651C5.388,8.474 7.533,2.959 12,1.5m0,0c4.468,1.46 6.612,6.974 7.317,9.849a0.856,0.856 0,0 0,0.833 0.651h0.4a1.796,1.796 0,0 0,1.908 -2.084C21.49,5.7 16.595,1.5 12,1.5" />
     <path
         android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M12.237,7.688c3.29,1.197 5.7,5.48 5.294,8.75 -0.109,0.878 -1.061,1.304 -1.908,0.995l-0.288,-0.104a0.65,0.65 0,0 1,-0.426 -0.684c0.245,-2.242 0.146,-6.748 -2.672,-8.957m-9.681,3.3c1.792,-2.767 6.391,-4.498 9.681,-3.3 -3.579,-0.12 -6.55,3.27 -7.804,5.144a0.65,0.65 0,0 1,-0.766 0.25l-0.288,-0.105c-0.847,-0.308 -1.304,-1.247 -0.823,-1.99m9.611,-4.76A3.751,3.751 0,0 1,18.99 9.1a3.73,3.73 0,0 1,-1.22 1.971c-1.044,-2.111 -2.928,-3.875 -5.396,-4.774zM21.034,8.63 L21.944,8.852a0.937,0.937 0,0 1,-0.442 1.822l-0.91,-0.22a0.938,0.938 0,0 1,0.442 -1.823M20.46,4a0.939,0.939 0,0 1,0.976 1.601l-0.8,0.488a0.938,0.938 0,0 1,-0.976 -1.6zM11.44,1.814a0.936,0.936 0,0 1,1.287 0.313l0.49,0.8a0.938,0.938 0,0 1,-1.6 0.977l-0.489,-0.8a0.94,0.94 0,0 1,0.312 -1.29m5.67,-0.887c0.504,0.122 0.812,0.63 0.69,1.133l-0.219,0.91a0.937,0.937 0,0 1,-1.822 -0.44l0.22,-0.913a0.94,0.94 0,0 1,1.132 -0.69" />
+        android:pathData="M12,1.5c-1.445,1.416 -4.807,6.26 -5.258,9.454A0.94,0.94 0,0 0,7.714 12h8.572a0.94,0.94 0,0 0,0.972 -1.046C16.808,7.761 13.445,2.916 12,1.5m0,12a1.5,1.5 0,0 0,-1.5 1.5v6a1.5,1.5 0,0 0,3 0v-6a1.5,1.5 0,0 0,-1.5 -1.5" />
 </vector>
