--- backpack-common/src/main/res/drawable-nodpi/bpk_temperature.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M10,7a1,1 0,0 1,1 1v4.82a0.56,0.56 0,0 0,0.325 0.492A3,3 0,0 1,13 16a3,3 0,1 1,-4.325 -2.688A0.56,0.56 0,0 0,9 12.82V8a1,1 0,0 1,1 -1" />
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M10,0a5,5 0,0 1,5 5v5.696c0,0.262 0.105,0.511 0.277,0.708a7,7 0,1 1,-10.555 0c0.173,-0.197 0.278,-0.446 0.278,-0.708V5a5,5 0,0 1,5 -5m0,2a3,3 0,0 0,-3 3v6.53a1,1 0,0 1,-0.333 0.743,5 5,0 1,0 6.666,0A1,1 0,0 1,13 11.53V5a3,3 0,0 0,-3 -3" />
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M20,8a1,1 0,1 1,0 2h-3L17,8zM20,5a1,1 0,1 1,0 2h-3L17,5zM20,2a1,1 0,1 1,0 2h-3L17,2z" />
-</vector>
