--- backpack-common/src/main/res/drawable-nodpi/bpk_temperature_sm.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:width="24dp"
-    android:height="24dp"
-    android:viewportWidth="24"
-    android:viewportHeight="24">
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M9,7.5A1.5,1.5 0,0 1,10.5 9v4.904a3,3 0,1 1,-3 0V9A1.5,1.5 0,0 1,9 7.5" />
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M9,0a6,6 0,0 1,6 6v6c0,0.75 1.5,2.812 1.5,4.5a7.5,7.5 0,0 1,-15 0c0,-1.688 0.75,-3 1.5,-4.5V6a6,6 0,0 1,6 -6m0,3a3,3 0,0 0,-3 3s0.106,6.405 0,6.75 -0.037,0.3 -0.602,1.052A4.5,4.5 0,1 0,13.5 16.5c0,-1.016 -0.49,-1.946 -1.055,-2.698 0,0 -0.445,-0.661 -0.445,-1.052V6a3,3 0,0 0,-3 -3" />
-    <path
-        android:fillColor="@color/bpkTextPrimary"
-        android:pathData="M21,7.5a1.5,1.5 0,0 1,0 3h-3v-3zM21,3a1.5,1.5 0,0 1,0 3h-3V3z" />
-</vector>
