+++ backpack-common/src/main/res/values-night/backpack.internal.color.xml
