+++ backpack-common/src/main/res/values-night/backpack.semantic.color.xml
