+++ backpack-common/src/main/res/values/backpack.animation.xml
