+++ backpack-common/src/main/res/values/backpack.internal.color.xml
