+++ backpack-common/src/main/res/values/backpack.radii.xml
@@ -20,11 +20,9 @@
 
   <!-- THIS FILE IS AUTO GENERATED. DO NOT MODIFY! -->
 
-    <dimen name="bpkBorderRadiusFull">9999dp</dimen>
-    <dimen name="bpkBorderRadiusXl">40dp</dimen>
-    <dimen name="bpkBorderRadiusMd">12dp</dimen>
-    <dimen name="bpkBorderRadiusNavTabs">18dp</dimen>
-    <dimen name="bpkBorderRadiusSm">8dp</dimen>
     <dimen name="bpkBorderRadiusXs">4dp</dimen>
+    <dimen name="bpkBorderRadiusSm">8dp</dimen>
+    <dimen name="bpkBorderRadiusMd">12dp</dimen>
     <dimen name="bpkBorderRadiusLg">24dp</dimen>
+    <dimen name="bpkBorderRadiusXl">40dp</dimen>
 </resources>
