+++ backpack-common/src/main/res/values/backpack.text.xml
@@ -27,13 +27,6 @@
       <attr name="android:letterSpacing" />
       <attr name="lineHeight" />
   </declare-styleable>
-  
-  <style name="bpkTextBaseLarken">
-    <item name="android:fontFamily">?bpkFontFamilyEmphasized</item>
-    <item name="android:textColor">@color/bpkTextPrimary</item>
-    <item name="android:textSize">@dimen/bpkTextBaseSize</item>
-    <item name="lineHeight">20sp</item>
-  </style>
 
   <style name="bpkTextBodyDefault">
     <item name="android:fontFamily">?bpkFontFamilyBase</item>
@@ -131,7 +124,7 @@
   </style>
 
   <style name="bpkTextHero5">
-    <item name="android:fontFamily">?bpkFontFamilyBlack</item>
+      <item name="android:fontFamily">?bpkFontFamilyEmphasized</item>
     <item name="android:textColor">@color/bpkTextPrimary</item>
     <item name="android:textSize">@dimen/bpkTextXxxxlSize</item>
     <item name="lineHeight">56sp</item>
@@ -159,24 +152,10 @@
     <item name="lineHeight">16sp</item>
   </style>
 
-  <style name="bpkTextSmLarken">
-    <item name="android:fontFamily">?bpkFontFamilyEmphasized</item>
-    <item name="android:textColor">@color/bpkTextPrimary</item>
-    <item name="android:textSize">@dimen/bpkTextSmSize</item>
-    <item name="lineHeight">20sp</item>
-  </style>
-
   <style name="bpkTextSubheading">
     <item name="android:fontFamily">?bpkFontFamilyBase</item>
     <item name="android:textColor">@color/bpkTextPrimary</item>
     <item name="android:textSize">@dimen/bpkTextXlSize</item>
     <item name="lineHeight">32sp</item>
-  </style>
-
-  <style name="bpkTextXsLarken">
-    <item name="android:fontFamily">?bpkFontFamilyEmphasized</item>
-    <item name="android:textColor">@color/bpkTextPrimary</item>
-    <item name="android:textSize">@dimen/bpkTextXsSize</item>
-    <item name="lineHeight">16sp</item>
   </style>
 </resources>
